+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - New feed event subscription
+- New fields for proposal event type 
 
 ## [0.1.6] - 2024-12-23